Renamed `collectstatic_storeys.py` to `collect_storeys_routes.py`.
